Move date formatter into its own module
