+++ src/shared/helpers/formatters/dateFormatter.ts
@@ -1,0 +1,32 @@
+/**
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export const getFormattedDate = (date: Date) => {
+  const year = date.getUTCFullYear();
+  const monthAsNumber = date.getUTCMonth() + 1; // month returned from date.getMonth() is zero-based
+  const month = String(monthAsNumber).padStart(2, '0');
+  const day = String(date.getUTCDate()).padStart(2, '0');
+
+  return `${year}-${month}-${day}`;
+};
+
+export const getFormattedDateTime = (date: Date) => {
+  const formattedDate = getFormattedDate(date);
+  const hours = String(date.getUTCHours()).padStart(2, '0');
+  const minutes = String(date.getUTCMinutes()).padStart(2, '0');
+
+  return `${formattedDate}, ${hours}:${minutes}`;
+};
